Redesign API
Given a git function that can return 'Either GitError a', we want to
allow multiple behaviors:

1. Returning the typed error, as-is.
2. Replacing the GitError with some default value of type 'a'.
3. Calling 'error' on the 'GitError', thus requiring success at
   build time.
4. Falling back to an environment variable lookup, when git fails.

Previously, we attempted to provide all combinations ourselves. But
there are several problems with this:

1. It is unwieldy.
2. Unclear if each behavior makes sense for all git functions
   (e.g. should we perform an env lookup for describe?)
3. The untyped API becomes very error prone.

We therefore remove the manual combinations, instead supplying the
lower-level combinators and emphasizing easy composition.

The untyped API remains the same as it used to be: If git fails,
some default value is returned.

The typed API, on the other hand, provides the same functionality
as the untyped one, but additionally it provides combinators
for each of the above behaviors. The user can therefore choose exactly
the functionality they want for a given function.