Redo env logic
Instead of 'fallback' functions, we make Env lookup functions
return git values as normal. This enables more flexibility as
it allows one to customize the order in which actions are run.

To facilitate this, we introduce the QFirst newtype that allows
one to easily combine actions via its Semigroup interface,
while also preserving laziness.